LPS-63709 Fix split package introduce in LPS-63360
