+++ modules/apps/collaboration/wiki/wiki-service/bnd.bnd
@@ -16,6 +16,7 @@ Export-Package:\
 	com.liferay.wiki.trash,\
 	com.liferay.wiki.util,\
 	com.liferay.wiki.util.comparator,\
+	com.liferay.wiki.validator,\
 	com.liferay.wiki.verify
 Import-Package:\
 	!com.liferay.portal.tools.sql,\
